+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ### Features
 
-- Introducing `@sentry/react-native/playground` ([#4916](https://github.com/getsentry/sentry-react-native/pull/4916), [#4918](https://github.com/getsentry/sentry-react-native/pull/4918))
+- Introducing `@sentry/react-native/playground` ([#4916](https://github.com/getsentry/sentry-react-native/pull/4916), [#4918](https://github.com/getsentry/sentry-react-native/pull/4918)))
 
   The new `withSentryPlayground` component allows developers to verify
   that the SDK is properly configured and reports errors as expected.